presentation/scripts/requirements.txt
python-pptx>=0.6.21
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"
pywin32>=306; sys_platform == "win32"
psutil>=5.9.0